backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
transformers==4.44.2
torch>=2.2.0
python-multipart==0.0.9
slowapi==0.1.9
